pyproject.toml
[project]
name = "aws-s3-migration"
version = "1.0.0"
description = "AWS S3 bucket migration and policy management tools"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "pytest>=7.4.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

# =============================================================================
# Repository-Specific Tool Configurations
# =============================================================================
# These are repo-specific and not standardized across repositories
# =============================================================================

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
skip = ["ci_shared"]

[tool.pylint.MASTER]
init-hook = "import sys, os; sys.path.insert(0, os.getcwd())"
ignored-modules = ["psycopg2"]

[tool.pylint.'MESSAGES CONTROL']
disable = []

[tool.pylint.DESIGN]
max-attributes = 10
max-args = 7

[tool.vulture]
exclude = ["ci_shared/", "scripts/", "docs/", ".venv/"]

[tool.deptry]
extend_exclude = ["ci_shared", "tests", "scripts"]
pep621_dev_dependency_groups = []
requirements_files_dev = ["scripts/requirements.txt"]
per_rule_ignores = { DEP001 = ["overview", "s3_audit", "billing_report", "ebs_manager", "rds_aurora_migration", "snapshot_export", "snapshot_export_fixed", "snapshot_export_robust", "analysis", "deletion", "workflow", "cache", "cleanup_temp_artifacts", "duplicate_tree", "find_compressible"], DEP002 = ["boto3", "pydantic"], DEP003 = ["ci_tools"] }

[tool.deptry.package_module_name_map]
"psycopg2-binary" = ["psycopg2"]

[tool.bandit]
exclude_dirs = [
    "/vendor/",
    "/.venv/",
    "/venv/",
    "/artifacts/",
    "/trash/",
    "/models/",
    "/logs/",
]
skips = [
    "B101",
    "B301",
    "B403",
    "B105",
    "B106",
    "B107",
    "B108",
    "B110",
    "B311",
    "B404",
    "B603",
    "B607",
    "B608",
]

[tool.coverage]

[tool.coverage.run]
omit = [
    "*/vendor/*",
    "*/__main__.py",
]

[tool.pytest]

[tool.pytest.ini_options]
addopts = [
    "--tb=short",
    "-q",
]
pythonpath = [
    ".",
]

[tool.ruff]
extend-exclude = [
    "ci_tools/vendor",
]

[tool.ruff.lint]
ignore = []
select = [
    "TRY",
    "C90",
    "PLR",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "TRY002",
    "TRY003",
    "TRY203",
    "PLR2004",
]

[tool.ruff.lint.pylint]
max-args = 7
max-branches = 10
max-statements = 50
